+ IsRoshan
* Moved Debug functions to a new header
