+++ Dota2Cheat/UIState.h
@@ -6,6 +6,7 @@ namespace UIState {
 	namespace Menu {
 
 	}
+
 	inline const char* WeatherList[] = { "Default", "Winter", "Rain", "MoonBeam", "Pestilence", "Harvest", "Sirocco", "Spring", "Ash", "Aurora" };
 
 	inline bool HeroVisibleToEnemy = false;
